# improve sign() performance by caching SigningCommitments