Closes https://github.com/ZcashFoundation/frost/issues/454

Benchmark:

```
FROST Signing ristretto255/Round 2/2
                        time:   [97.239 µs 97.456 µs 97.727 µs]
                        change: [-28.248% -28.139% -28.010%] (p = 0.00 < 0.05)
                        Performance has improved.
FROST Signing ristretto255/Round 2/7
                        time:   [180.00 µs 181.16 µs 182.72 µs]
                        change: [-19.124% -18.843% -18.525%] (p = 0.00 < 0.05)
                        Performance has improved.
FROST Signing ristretto255/Round 2/67
                        time:   [1.1946 ms 1.1955 ms 1.1966 ms]
                        change: [-3.4950% -3.2638% -3.0433%] (p = 0.00 < 0.05)
                        Performance has improved.
FROST Signing ristretto255/Round 2/667
                        time:   [11.588 ms 11.607 ms 11.629 ms]
                        change: [-1.4843% -1.2623% -1.0580%] (p = 0.00 < 0.05)
                        Performance has improved.
```